Revert "Restore health sensors after restart"
