+++ custom_components/unifi_gateway_refactored/diagnostics.py
@@ -5,7 +5,6 @@ from typing import Any, Dict
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
-from homeassistant.components.diagnostics import async_redact_data
 
 from .const import (
     CONF_HOST,
@@ -26,56 +25,6 @@ from .const import (
 from .unifi_client import UniFiOSClient
 
 
-TO_REDACT: set[str] = {
-    CONF_HOST,
-    CONF_USERNAME,
-    CONF_PASSWORD,
-    CONF_SITE_ID,
-    "controller_ui",
-    "controller_api",
-    "site",
-    "mac",
-    "mac_address",
-    "macs",
-    "serial",
-    "device_id",
-    "ip",
-    "ip_address",
-    "ipv4",
-    "ipv6",
-    "wan_ip",
-    "lan_ip",
-    "gateway",
-    "gateway_ip",
-    "ssid",
-    "essid",
-    "bssid",
-    "token",
-    "cookie",
-    "secret",
-    "passphrase",
-    "password",
-    "username",
-    "private_key",
-    "key",
-    "session_id",
-}
-
-
-def _build_diagnostics_payload(client: UniFiOSClient) -> Dict[str, Any]:
-    """Collect controller diagnostics payload from the UniFi client."""
-
-    health = client.get_healthinfo()
-    sites = client.list_sites()
-    return {
-        "controller_ui": client.get_controller_url(),
-        "controller_api": client.get_controller_api_url(),
-        "site": client.get_site(),
-        "health": health,
-        "sites": sites,
-    }
-
-
 async def async_get_config_entry_diagnostics(
     hass: HomeAssistant, entry: ConfigEntry
 ) -> Dict[str, Any]:
@@ -85,10 +34,17 @@ async def async_get_config_entry_diagnostics(
         client: UniFiOSClient = stored["client"]
 
         def _collect_existing() -> Dict[str, Any]:
-            return _build_diagnostics_payload(client)
+            health = client.get_healthinfo()
+            sites = client.list_sites()
+            return {
+                "controller_ui": client.get_controller_url(),
+                "controller_api": client.get_controller_api_url(),
+                "site": client.get_site(),
+                "health": health,
+                "sites": sites,
+            }
 
-        payload = await hass.async_add_executor_job(_collect_existing)
-        return async_redact_data(payload, TO_REDACT)
+        return await hass.async_add_executor_job(_collect_existing)
 
     def _collect_fresh() -> Dict[str, Any]:
         client = UniFiOSClient(
@@ -103,7 +59,14 @@ async def async_get_config_entry_diagnostics(
             ),
             timeout=entry.data.get(CONF_TIMEOUT, DEFAULT_TIMEOUT),
         )
-        return _build_diagnostics_payload(client)
+        health = client.get_healthinfo()
+        sites = client.list_sites()
+        return {
+            "controller_ui": client.get_controller_url(),
+            "controller_api": client.get_controller_api_url(),
+            "site": client.get_site(),
+            "health": health,
+            "sites": sites,
+        }
 
-    payload = await hass.async_add_executor_job(_collect_fresh)
-    return async_redact_data(payload, TO_REDACT)
+    return await hass.async_add_executor_job(_collect_fresh)
